Core_software/requirements.txt
pypdf>=4.0,<5
tkinterdnd2>=0.4.2,<0.5
customtkinter>=5.2.2,<6
Pillow>=10,<13
rlottie-python>=1.3.8,<2
PyMuPDF>=1.24,<2
PySide6>=6.6,<7

# Windows setup (run in PowerShell):
# 1) Install Python (if needed):
#    winget install --id Python.Python.3.13 -e --accept-package-agreements --accept-source-agreements
# 2) Install SumatraPDF (required for printing):
#    winget install --id SumatraPDF.SumatraPDF -e --accept-package-agreements --accept-source-agreements
# 3) Install Python dependencies:
#    python -m pip install -r requirements.txt
# 4) (Optional) Verify SumatraPDF is on PATH:
#    where.exe SumatraPDF.exe
#
# Windows setup (run in PowerShell):
# 1) Install Python (if needed):
#    winget install --id Python.Python.3.13 -e --accept-package-agreements --accept-source-agreements
# 2) Install SumatraPDF (required for printing):
#    winget install --id SumatraPDF.SumatraPDF -e --accept-package-agreements --accept-source-agreements
# 3) (Optional) Verify SumatraPDF is on PATH:
#    where.exe SumatraPDF.exe
